+++ Sources/HIGThemesSystem/HIGSystemTheme.swift
@@ -39,6 +39,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
     public let ribbon: any HIGRibbonTokens
     public let pricingCard: any HIGPricingCardTokens
     public let chatBubble: any HIGChatBubbleTokens
+    public let comment: any HIGCommentTokens
     public let dataTable: any HIGDataTableTokens
     public let breadcrumb: any HIGBreadcrumbTokens
     public let pageHeader: any HIGPageHeaderTokens
@@ -126,6 +127,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
         ribbon: any HIGRibbonTokens = HIGSystemRibbonTokens(),
         pricingCard: any HIGPricingCardTokens = HIGSystemPricingCardTokens(),
         chatBubble: any HIGChatBubbleTokens = HIGSystemChatBubbleTokens(),
+        comment: any HIGCommentTokens = HIGSystemCommentTokens(),
         dataTable: any HIGDataTableTokens = HIGSystemDataTableTokens(),
         breadcrumb: any HIGBreadcrumbTokens = HIGSystemBreadcrumbTokens(),
         pageHeader: any HIGPageHeaderTokens = HIGSystemPageHeaderTokens(),
@@ -212,6 +214,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
         self.ribbon = ribbon
         self.pricingCard = pricingCard
         self.chatBubble = chatBubble
+        self.comment = comment
         self.dataTable = dataTable
         self.breadcrumb = breadcrumb
         self.pageHeader = pageHeader